Changed the ProgramVersion to 3.3.2 and increased its size
I've reviewed the uses of ProgramVersion; my conclusion is

a. The size of the variable *should* be 10 chars rather than 8
b. Changing it should be harmless
c. I suspect the code will now read older files correctly

Note that the variable is used in licensing: 3.3.2 will be backwards
compatible with 3.3 licenses.